Move Executor to separate module
